fix some C140/C219 issues
- fixed crash with C219 data block (odd number of bytes)
- fixed crackling in C140 (precision loss during interpolation)
- fixed memory corruption in C219 (trying to access voices memory 16..23)

+++ emu/cores/c140.c
@@ -322,7 +322,7 @@ static void c140_update(void *param, UINT32 samples, DEV_SMPL **outputs)
 					}
 
 					/* Caclulate the sample value */
-					dt=((v->dltdt*v->ptoffset)>>16)+v->prevdt;
+					dt=(INT32)(((INT64)v->dltdt*v->ptoffset)>>16)+v->prevdt;
 
 					/* Write the data to the sample buffers */
 					lmix[j]+=(dt*vreg->volume_left)>>8;
@@ -361,7 +361,7 @@ static void c140_update(void *param, UINT32 samples, DEV_SMPL **outputs)
 					}
 
 					/* Caclulate the sample value */
-					dt=((v->dltdt*v->ptoffset)>>16)+v->prevdt;
+					dt=(INT32)(((INT64)v->dltdt*v->ptoffset)>>16)+v->prevdt;
 
 					/* Write the data to the sample buffers */
 					lmix[j]+=(dt*vreg->volume_left)>>8;
